Refactor: move to per-feature package model.
